fix: issue for overflowing row items while formatting
